+++ src/main/java/Controller/HomeController.java
@@ -15,10 +15,6 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-/**
- *
- * @author HP
- */
 public class HomeController extends HttpServlet {
 
     /**
